initial upload of Will's obsplan code
The purpose of this program is to help with planning the DEIMOS observations, in
particular the galaxy selection part of the mask making.

Note that while this program can handle multiple mask regions at one time,
the output deimos input will not be meaningful.

Version 0 of this code is based on obsplan.py that was used in the planning of
the Musket Ball 2011A DEIMOS run.