corrected obsplan.objectPA() function
I corrected this function so that it calculates the correct object parallactic angle
for an observer in the northern hemisphere.  Before the calculation was not adding
180 degrees to q if the denominator was negative.  I have verified that this function
works correctly see debug_optimalPA and compare palomar_pa.txt with Table II of
Filippenko (1982).

I also deleted the redundant code in optimalPA and will turn this into a DEIMOS
slit mask specific function.

Note that this has not been verified for observers in the southern hemisphere.