Replaced old make_bintab_file.pro with latest version edited by Brian Lemaux
Brian edited the code to fix a problem where the extraction window would
be mirrored on the slight hight and thus extracting from the wrong end of
the slit. This fix should be general but it has not been thoroughly tested yet.

Dear Nathan and William,

I have fixed your problem. At least as it relates to the mask that you sent me. I believe the solution will work in general though. The problem... I dun feel like explaining it now, but I can explain it to you later if you're really interested. The solution: take the attached file, place it in your spec2d/pro/ directory, delete everything having to do with your previous reductions for the masks with problems (including the zresult files), and re-reduce the masks that had this problem using the exact same calling sequence as before. After that, tell me if it worked or not.

IPAz please.

-B